restructured tests
