+++ tests/unit/Models/LicenseTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CycloneDX\Tests\Models;
+namespace CycloneDX\Tests\uni\Models;
 
 use CycloneDX\Models\License;
 use PHPUnit\Framework\TestCase;
